new sugestions work in progress
